feat: Add user-agent check for server_data endpoint
This commit adds a user-agent check to the `/growtopia/server_data.php` endpoint in order to restrict access to requests made by the "UbiServices_SDK" user agent. If the user-agent header is empty or does not contain "UbiServices_SDK", a 403 Forbidden response is sent. This helps ensure that only authorized clients can access the server data endpoint.